requirements.txt
PyQt6
pdf2image
Pillow